+++ bb/switch_tex_styles.clj
@@ -9,11 +9,11 @@
 
 (defn usage
   []
-  (println "switch-tex-styles version" version)
+  (println "switch-tex-styles, version" version)
   (println "usage: switch-tex-styles {--gaku|--my} dir")
-  (println "  ファイル dir/*.tex の句読点、全角半角間のスペースを調整する。")
-  (println "  --gaku 句読点は 「.,」を使い、全角半角間のスペースを削除。")
-  (println "  --my 句読点は 「。、」を使い、全角半角間にスペースを入れる。"))
+  (println "ファイル dir/*.tex の句読点、全角半角間のスペースを調整する。")
+  (println " --gaku ... 句読点は 「．，」を使い、全角半角間のスペースを削除。")
+  (println " --my ..... 句読点は 「。、」を使い、全角半角間にスペースを入れる。"))
 
 (defn tex-files
   "tex files in dir."